* 'develop' of https://github.com/iOS-EST-2nd-LeaveGo/LeaveGo: (24 commits)
  fix: 여행썸네일 삭제시 기본이미지 제거
  fix: 이미지 캐싱 처리
  feat: ImageCacheManager 추가
  feat: 상세페이지 설정
  design: 지도 헤더 UI 간격 조정, 장소 목록에서 운영시간 라벨 숨기기
  design: 지도 헤더 UI 간격 조정
  design: 여행 추가 화면 레이아웃 수정, 교통수단 수정 ( 자전거 → 대중교통 )
  feat: createPlannerPlace
  design: 온보딩뷰 디자인 수정
  feat: 여행 썸네일 표기, 여행 삭제 기능 구현
  feat: 길게 눌러 삭제하기 분기 처리 - 1트 실패
  design: 플레이스홀더 이미지 이름 변경 (#88)
  design: 플레이스홀더 이미지 이름 변경
  chore: 테스트 끝난 print 제거
  feat : 전달 받은 ID 가 있다면 여행만들기 버튼 숨김
  [fix] user annotation 이미지 테두리가 좀 더 하얀 이미지로 변경, alert 문구 수정
  fix: LocationManager 사용하면서 발생한 List 문제 발생
  docs: 불필요한 파일 제거
  feat: fetchOnePlanner 적용 -> 상세페이지 여행이름, 썸네일 랜더링
  design: 컬러셋 수정, 각종 플레이스홀더 이미지 적용
  ...

# Conflicts:
#	LeaveGo/Feature/Places/MapViewController.swift

+++ LeaveGo/Shared/Persistence/Coredata/CoreDataManager+Planner.swift
@@ -72,8 +72,9 @@ extension CoreDataManager {
         planner.thumbnailPath = thumbnailPath
         saveContext()
     }
-
-    func deletePlanner(_ planner: PlannerEntity) {
+    
+    func deletePlanner(id: UUID) {
+        guard let planner = fetchOnePlanner(id: id) else { return }
         context.delete(planner)
         saveContext()
     }
